add logic to consolidate aggregate row data
