+++ services/QuillLMS/app/assets/stylesheets/pages/premium_hub/growth-diagnostic-reports.scss
@@ -67,6 +67,8 @@ main {
     margin-top: 32px;
     max-width: 300px;
     width: 100%;
+    height: 60px;
+    justify-content: flex-start;
   }
   .growth-diagnostic-reports-overview-table, .growth-diagnostic-reports-by-skill-table {
     margin-top: 48px;
